services/api-gateway/requirements.txt
# API Gateway Requirements

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for service communication
httpx==0.25.2
aiohttp==3.9.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# JWT handling
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Rate limiting and caching
slowapi==0.1.9
aioredis==2.0.1

# Circuit breaker
circuitbreaker==1.4.0

# Load balancing and service discovery
consul==1.1.0

# Environment variables
python-dotenv==1.0.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0